requirements.txt
argon2-cffi==21.1.0
attrs==21.2.0
backcall==0.2.0
bleach==4.1.0
cffi==1.15.0
click==8.0.4
colorama==0.4.4
cycler==0.11.0
Cython==0.29.28
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
entrypoints==0.3
Flask==2.0.1
gunicorn==20.1.0
ipykernel==6.4.2
ipython==7.29.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==2.1.0
jedi==0.18.0
Jinja2==3.0.2
jsonschema==4.1.2
jupyter==1.0.0
jupyter-client==7.0.6
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
mistune==0.8.4
mysql-connector-python==8.0.28
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.5
numpy==1.21.3
packaging==21.3
pandas==1.3.4
pandocfilters==1.5.0
parso==0.8.2
patsy==0.5.2
pickleshare==0.7.5
Pillow==8.4.0
prometheus-client==0.12.0
prompt-toolkit==3.0.21
protobuf==3.19.4
pycparser==2.20
Pygments==2.10.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.3
pyzmq==22.3.0
qtconsole==5.1.1
QtPy==1.11.2
scipy==1.8.0
Send2Trash==1.8.0
six==1.16.0
statsmodels==0.13.2
terminado==0.12.1
testpath==0.5.0
tornado==6.1
traitlets==5.1.1
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.3
widgetsnbextension==3.5.2
flask-cors==3.0.10
furl==2.1.3